requirements.txt
python-dotenv
qdrant-client
nltk
sentence-transformers
flask
openai
